refactor/#336 서비스타입 변경사항 반영
